Fix for PDI breaking changes
- xyz.openbmc_project.Inventory.Item.Dimm MemorySizeInKB type was
  changed from uint32 to size.
- xyz.openbmc_project.Common.UUID UUID field was renamed (in the
  generated server binding code) from uUID to uuid. Use the sdbus++
  preprocessor define to tell if this name change is in effect or not,
  in order to build with multiple versions of sdbusplus.

Tested: Build passes.

+++ src/dimm.cpp
@@ -101,7 +101,7 @@ void Dimm::dimmSizeExt(uint32_t size)
     memorySizeInKB(size);
 }
 
-uint32_t Dimm::memorySizeInKB(uint32_t value)
+size_t Dimm::memorySizeInKB(size_t value)
 {
     return sdbusplus::xyz::openbmc_project::Inventory::Item::server::Dimm::
         memorySizeInKB(value);
